requirements.txt
openai>=1.0.0
pyaudio>=0.2.11
pyttsx3>=2.90
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
numpy>=1.24.0
sounddevice>=0.4.6
scipy>=1.11.0
langchain>=0.1.0
langchain-openai>=0.1.0
pydantic>=2.0.0
pyinstaller>=5.0.0
pystray>=0.19.0
pillow>=9.0.0
python-docx>=0.8.11
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
ultralytics>=8.0.0 